+++ test/baselines/adapter_JS_ESLint_TypeChecking_Spaces_SingleQuotes_LGPLv3/admin/words.js
@@ -2,7 +2,6 @@
 'use strict';
 
 systemDictionary = {
-
     'test-adapter adapter settings': {
         'en': 'Adapter settings for test-adapter',
         'de': 'Adaptereinstellungen für test-adapter',
@@ -15,29 +14,28 @@ systemDictionary = {
         'pl': 'Ustawienia adaptera dla test-adapter',
         'zh-cn': 'test-adapter的适配器设置'
     },
-    'option 1 description': {
-        'en': 'Option 1 is cool',
-        'de': 'Option 1 ist cool',
-        'ru': 'Вариант 1 - это круто',
-        'pt': 'Opção 1 é legal',
-        'nl': 'Optie 1 is cool',
-        'fr': "L'option 1 est cool",
-        'it': "L'opzione 1 è cool",
-        'es': 'La opción 1 es genial',
-        'pl': 'Opcja 1 jest fajna',
-        'zh-cn': '选项1很酷'
+    'option1': {
+        'en': "Mock translation of 'option1' to 'en'",
+        'de': "Mock translation of 'option1' to 'de'",
+        'ru': "Mock translation of 'option1' to 'ru'",
+        'pt': "Mock translation of 'option1' to 'pt'",
+        'nl': "Mock translation of 'option1' to 'nl'",
+        'fr': "Mock translation of 'option1' to 'fr'",
+        'it': "Mock translation of 'option1' to 'it'",
+        'es': "Mock translation of 'option1' to 'es'",
+        'pl': "Mock translation of 'option1' to 'pl'",
+        'zh-cn': "Mock translation of 'option1' to 'zh-cn'"
     },
-    'option 2 description': {
-        'en': 'Option 2 is not',
-        'de': 'Option 2 nicht',
-        'ru': 'Вариант 2 не',
-        'pt': 'A opção 2 não é',
-        'nl': 'Optie 2 is dat niet',
-        'fr': "L'option 2 n'est pas",
-        'it': "L'opzione 2 no",
-        'es': 'La opción 2 no es',
-        'pl': 'Opcja 2 nie jest',
-        'zh-cn': '选项2不是'
+    'option2': {
+        'en': "Mock translation of 'option2' to 'en'",
+        'de': "Mock translation of 'option2' to 'de'",
+        'ru': "Mock translation of 'option2' to 'ru'",
+        'pt': "Mock translation of 'option2' to 'pt'",
+        'nl': "Mock translation of 'option2' to 'nl'",
+        'fr': "Mock translation of 'option2' to 'fr'",
+        'it': "Mock translation of 'option2' to 'it'",
+        'es': "Mock translation of 'option2' to 'es'",
+        'pl': "Mock translation of 'option2' to 'pl'",
+        'zh-cn': "Mock translation of 'option2' to 'zh-cn'"
     }
-
 };